# apps/api/api/openapi.yaml
openapi: 3.1.0
info:
  title: TypeCall API
  version: 1.0.0
  description: TypeCall — Typeform + Calendly fusion. Qualification + scheduling in one conversational flow.
  contact:
    name: milennials
    email: fabiomilennials1234@gmail.com
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.typecall.io
    description: Production
...
paths:
...
  /api/v1/public/events:
    post:
      tags: [Public]
      operationId: ingestEvents
      summary: Ingest analytics events (max 50 per batch)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestEventsInput"
      responses:
        "200":
          description: Events ingested
          content:
            application/json:
              schema:
                type: object
                properties:
                  inserted:
                    type: integer
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
...
  schemas:
...
    AnalyticsEventType:
      type: string
      enum: [view, start, question_seen, question_answered, booking_slot_selected, submit, abandon, share]
    IngestEventsInput:
      type: object
      required: [events]
      properties:
        events:
          type: array
          maxItems: 50
          items:
            type: object
            required: [event_id, form_id, event_type]
            properties:
              event_id:
                type: string
                format: uuid
              form_id:
                type: string
                format: uuid
              response_id:
                type: string
                format: uuid
              step_id:
                type: string
              event_type:
                $ref: "#/components/schemas/AnalyticsEventType"
              metadata: {}

    # ─── Webhooks ─────────────────────────────────────────
...